Add options and tidy some tables/functions
The values for items such as seek or audio/video sync are hardcoded, so options are added to make them configurable. We also tidy some of the tables and functions and update the shorcut/accelerator handling to dynamically size the menus based on maximum shortcut length for a given menu.