Update README.md and fix Tk menu re-post
Updated the README.md with a few readability changes including a note about the Gtk+ menu opening delay. Also fixed an issue with the re-posting (postcascade) not using the x,y location that is provided back by the TCL script which caused the re-posted menu to keep moving, rather than staying in place.