Update to menu, builder and dialogs
Updates to the menu reflect the change to the dialogs script (now gui-dialogs.lua instead of zenity-dialogs.lua) which allows for use of both kdialog and zenity for dialogs. The gtk menu builder has also been updated to fix errors in relation to the version of Gdk/Gtk during imports.